+++ test/runtests.jl
@@ -247,6 +247,10 @@ end
         @test @inferred(tan(a))  ≈ sin(a)  / cos(a)
         @test @inferred(tand(a)) ≈ sind(a) / cosd(a)
         @test @inferred(tanh(a)) ≈ sinh(a) / cosh(a)
+        @test @inferred(sinpi(a))≈ sin(pi * a) atol = 1e-15
+        @test @inferred(cospi(a))≈ cos(pi * a) atol = 1e-15
+        @test @inferred(sinc(a)) ≈ (sin(pi * a) / (pi * a)) atol = 1e-16
+        @test @inferred(cosc(a)) ≈ ((cos(pi * a) - sin(pi * a) / (pi * a)) / a) atol = 1e-16
     end
 end